move the extension classes to arrow_canonical_extensions.py and fix fast/udf/test_scalar.py
