+++ tools/pythonpkg/tests/arrow_canonical_extensions.py
@@ -33,3 +33,25 @@ class JSONType(pa.ExtensionType):
         # return an instance of this subclass given the serialized
         # metadata.
         return JSONType()
+
+class UHugeIntType(pa.ExtensionType):
+    def __init__(self):
+        pa.ExtensionType.__init__(self, pa.binary(16), "duckdb.uhugeint")
+
+    def __arrow_ext_serialize__(self):
+        return b''
+
+    @classmethod
+    def __arrow_ext_deserialize__(self, storage_type, serialized):
+        return UHugeIntType()
+
+class HugeIntType(pa.ExtensionType):
+    def __init__(self):
+        pa.ExtensionType.__init__(self, pa.binary(16), "duckdb.hugeint")
+
+    def __arrow_ext_serialize__(self):
+        return b''
+
+    @classmethod
+    def __arrow_ext_deserialize__(self, storage_type, serialized):
+        return HugeIntType()
